[FLINK-15090][tests] Move streaming env tests to flink-tests
+ GetOperatorUniqueIDTest
+ RemoteStreamEnvironmentTest
+ ReinterpretDataStreamAsKeyedStreamITCase
+ LocalStreamEnvironmentITCase
